% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_criteria.R
\name{criteria_func}
\alias{criteria_func}
\title{Compute the four model-selection criteria indexes}
\usage{
criteria_func(Sigma, m, n, p)
}
\arguments{
\item{Sigma}{MLE estimated covariance matrix of error terms}

\item{m}{Dimension of multivariate time series}

\item{n}{Number of observations}

\item{p}{Order of estimated VAR(p)}
}
\value{
A list with (AIC,AIC_c,AIC_cBD,SIC).
}
\description{
Compute the four model-selection criteria indexes
}

