% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_criteria.R
\name{estimate_Delta}
\alias{estimate_Delta}
\title{Compute E[Delta(hat{beta},hat{Sigma})].}
\usage{
estimate_Delta(Phi, Sigma, Phi0, Sigma0, m, p, p0, n)
}
\arguments{
\item{Phi}{Estimated Phi, p-list with mxm transition matrices.}

\item{Sigma}{Estimated Sigma, mxm matrix with covariance matrix}

\item{Phi0}{True Phi, p-list with mxm transition matrices.}

\item{Sigma0}{True Sigma, mxm matrix with covariance matrix}

\item{m}{Number of dimensions of multivariate time series.}

\item{p}{Order of estimated VAR(p)}

\item{p0}{Order of true VAR(p0)}

\item{n}{Number of observations}
}
\description{
Compute E[Delta(hat{beta},hat{Sigma})].
}

