% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_param_vec}
\alias{get_param_vec}
\title{Convert beta and Sigma matrices into a vector of parameters}
\usage{
get_param_vec(beta, Sigma)
}
\arguments{
\item{beta}{Stacked matrix of transition matrices (Phi)}

\item{Sigma}{Covariance matrix of error terms}
}
\value{
Vector of parameters. The cells for Sigma correspond to the Cholesky
decomposition.
}
\description{
Convert beta and Sigma matrices into a vector of parameters
}

