% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_parameters}
\alias{get_parameters}
\title{Convert a vector of parameters into beta and Sigma matrices}
\usage{
get_parameters(param, p, m)
}
\arguments{
\item{param}{First $p*(m^2)$ parameters are beta. The rest
are Sigma. The parameters are the cells of the Cholesky decomposition.-}

\item{p}{Order of VAR(p)}

\item{m}{Number of dimensions of time series}
}
\description{
Convert a vector of parameters into beta and Sigma matrices
}

