% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lik_VAR}
\alias{lik_VAR}
\title{Negative log-likelihood}
\usage{
lik_VAR(param, Y, X, p, m, n)
}
\arguments{
\item{param}{Vector of parameters}

\item{Y}{m-variate time series}

\item{X}{lagged matrix of order (p). Does not include constant.}

\item{p}{Order of VAR(p)}

\item{m}{Dimension of time series}
}
\value{
Negative log-likelihood up to a constant
}
\description{
Negative log-likelihood
}

