% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mle_VAR}
\alias{mle_VAR}
\title{Obtain maximum likelihood estimate}
\usage{
mle_VAR(data, m, p, n, NumLags)
}
\arguments{
\item{data}{mx(n+NumLags) matrix}

\item{m}{Number of dimensions of time series}

\item{p}{Order of VAR(p)}

\item{NumLags}{Maximum number of lags set to zero in pre-data observations.}
}
\value{
List with (mp x m) matrix of betas and (mxm) matrix Sigma MLE
estiamtes.
}
\description{
Obtain maximum likelihood estimate
}

