% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{replicate_results}
\alias{replicate_results}
\title{Replicate the results in Hurvich and Tsai}
\usage{
replicate_results(Phi0, Sigma0, n, Nsim, m, p0, NumLags)
}
\arguments{
\item{Phi0}{A list with p transition matrices of a VAR(p)}

\item{Sigma0}{Covariance matrix of error terms}

\item{n}{Number of observations}

\item{Nsim}{Number of simulation draws}

\item{m}{Dimensions of multivariate time series}

\item{NumLags}{Maximum number of lag-order for evaluation}
}
\value{
A list of two tables summarizing the results. The first computes the
frequency with which each model order is selected. The second computes the
average
}
\description{
Replicate the results in Hurvich and Tsai
}
\examples{
do.call(replicate_results,example1(Nsim=100))
do.call(replicate_results,example2(Nsim=100))

}

